Adapt enrich tests and move them to AuthorizationTests
