+++ role-permissions-test.json
@@ -12,11 +12,171 @@
       }
     }
   ],
+  "view-columns": [
+    {
+      "type": "grant",
+      "action": ["view"],
+      "scope": "column"
+    }
+  ],
   "delete-tables": [
     {
       "type": "grant",
       "action": ["delete"],
       "scope": "table"
+    }
+  ],
+    "create-tables": [
+      {
+        "type": "grant",
+        "action": ["create"],
+        "scope": "table"
+      }
+    ],
+  "edit-tables": [
+    {
+      "type": "grant",
+      "action": ["view", "editDisplayProperty", "editStructureProperty"],
+      "scope": "table"
+    }
+  ],
+    "edit-model-tables": [
+      {
+        "type": "grant",
+        "action": ["view", "editDisplayProperty"],
+        "scope": "table"
+      },
+      {
+        "type": "grant",
+        "action": ["editStructureProperty"],
+        "scope": "table",
+        "condition": {
+          "table": {
+            "name": ".*_model"
+          }
+        }
+      }
+    ],
+      "edit-tables-all-allowed": [
+        {
+          "type": "grant",
+          "action": ["view", "editDisplayProperty", "editStructureProperty", "delete",
+            "createRow", "deleteRow", "editCellAnnotation", "editRowAnnotation"],
+          "scope": "table"
+        },
+        {
+          "type": "grant",
+          "action": ["view", "create"],
+          "scope": "column"
+        }
+      ],
+        "create-columns": [
+          {
+            "type": "grant",
+            "action": ["view"],
+            "scope": "table"
+          },
+          {
+            "type": "grant",
+            "action": ["view", "create"],
+            "scope": "column"
+          }
+        ],
+          "edit-columns-all-allowed": [
+            {
+              "type": "grant",
+              "action": ["view"],
+              "scope": "table"
+            },
+            {
+              "type": "grant",
+              "action": ["view", "editDisplayProperty", "editStructureProperty", "delete",
+                "editCellValue"],
+              "scope": "column"
+            }
+          ],
+            "can-edit-de-en-cell-values": [
+              {
+                "type": "grant",
+                "action": ["view"],
+                "scope": "table"
+              },
+              {
+                "type": "grant",
+                "action": ["view", "editCellValue"],
+                "scope": "column",
+                "condition": {
+                  "langtag": "en|es"
+                }
+              }
+            ],
+            "can-edit-cell-values": [
+              {
+                "type": "grant",
+                "action": ["view"],
+                "scope": "table"
+              },
+              {
+                "type": "grant",
+                "action": ["view", "editCellValue"],
+                "scope": "column"
+              }
+            ],
+             "can-edit-cell-values-AT-GB": [
+               {
+                 "type": "grant",
+                 "action": ["view"],
+                 "scope": "table"
+               },
+               {
+                 "type": "grant",
+                 "action": ["view", "editCellValue"],
+                 "scope": "column",
+                 "condition": {
+                   "langtag": "AT|GB"
+                 }
+               }
+             ],
+  "create-media": [
+    {
+      "type": "grant",
+      "action": ["create"],
+      "scope": "media"
+    }
+  ],
+    "edit-delete-media": [
+      {
+        "type": "grant",
+        "action": ["edit", "delete"],
+        "scope": "media"
+      }
+    ],
+  "NOT-delete-media": [
+    {
+      "type": "deny",
+      "action": ["delete"],
+      "scope": "media"
+    }
+  ],
+  "view-create-services": [
+    {
+      "type": "grant",
+      "action": ["view", "create"],
+      "scope": "service"
+    }
+  ],
+  "view-services": [
+    {
+      "type": "grant",
+      "action": ["view"],
+      "scope": "service"
+    }
+  ],
+  "edit-and-delete-services": [
+    {
+      "type": "grant",
+      "action": ["view", "create", "editDisplayProperty", "editStructureProperty", "delete"],
+      "scope": "service"
     }
   ],
   // Developer test role to be able to pass all integration tests
