+++ src/test/scala/com/campudus/tableaux/api/auth/MediaControllerAuthTest.scala
@@ -385,122 +385,3 @@ class MediaControllerAuthTest_checkAuthorization extends MediaControllerAuthTest
     }
 }
 
-@RunWith(classOf[VertxUnitRunner])
-class MediaControllerAuthTest_enrichAuthorization extends MediaControllerAuthTestBase {
-
-  @Test
-  def enrich_noPermission_noActionIsAllowed(implicit c: TestContext): Unit =
-    okTest {
-      val controller = createMediaController()
-
-      for {
-        permission <- controller.retrieveRootFolder("de").map(getPermission)
-      } yield {
-
-        val expected = Json.obj(
-          "create" -> false,
-          "edit" -> false,
-          "delete" -> false
-        )
-
-        assertJSONEquals(expected, permission)
-      }
-    }
-
-  @Test
-  def enrich_createGranted_onlyCreateIsAllowed(implicit c: TestContext): Unit = {
-    val roleModel = initRoleModel("""
-                                    |{
-                                    |  "create-media": [
-                                    |    {
-                                    |      "type": "grant",
-                                    |      "action": ["create"],
-                                    |      "scope": "media"
-                                    |    }
-                                    |  ]
-                                    |}""".stripMargin)
-
-    val controller = createMediaController(roleModel)
-
-    okTest {
-      for {
-        permission <- controller.retrieveRootFolder("de").map(getPermission)
-      } yield {
-
-        val expected = Json.obj("create" -> true)
-
-        assertJSONEquals(expected, permission, JSONCompareMode.LENIENT)
-      }
-    }
-  }
-
-  @Test
-  def enrich_editAndDeleteGranted_onlyEditAndDeleteAreAllowed(implicit c: TestContext): Unit = {
-    val roleModel = initRoleModel("""
-                                    |{
-                                    |  "edit-delete-media": [
-                                    |    {
-                                    |      "type": "grant",
-                                    |      "action": ["edit", "delete"],
-                                    |      "scope": "media"
-                                    |    }
-                                    |  ]
-                                    |}""".stripMargin)
-
-    val controller = createMediaController(roleModel)
-
-    okTest {
-      for {
-        permission <- controller.retrieveRootFolder("de").map(getPermission)
-      } yield {
-
-        val expected = Json.obj(
-          "create" -> false,
-          "edit" -> true,
-          "delete" -> true
-        )
-
-        assertJSONEquals(expected, permission)
-      }
-    }
-  }
-
-  @Test
-  def enrich_editAndDeleteGranted_deleteDenied_onlyEditIsAllowed(implicit c: TestContext): Unit = {
-    val roleModel = initRoleModel("""
-                                    |{
-                                    |  "edit-delete-media": [
-                                    |    {
-                                    |      "type": "grant",
-                                    |      "action": ["edit", "delete"],
-                                    |      "scope": "media"
-                                    |    }
-                                    |  ],
-                                    |  "NOT-delete-media": [
-                                    |    {
-                                    |      "type": "deny",
-                                    |      "action": ["delete"],
-                                    |      "scope": "media"
-                                    |    }
-                                    |  ]
-                                    |}""".stripMargin)
-
-    val controller = createMediaController(roleModel)
-
-    okTest {
-      for {
-        permission <- controller.retrieveRootFolder("de").map(getPermission)
-      } yield {
-
-        val expected = Json.obj(
-          "create" -> false,
-          "edit" -> true,
-          "delete" -> false
-        )
-
-        assertJSONEquals(expected, permission)
-      }
-    }
-  }
-
-}
