+++ src/test/scala/com/campudus/tableaux/api/auth/StructureControllerAuthTest.scala
@@ -1175,482 +1175,3 @@ class StructureControllerAuthTest_filterAuthorization extends StructureControlle
   }
 }
 
-@RunWith(classOf[VertxUnitRunner])
-class StructureControllerAuthTest_enrichDomainObjects extends StructureControllerAuthTest {
-
-  @Test
-  def enrichTableSeq_createIsAllowed(implicit c: TestContext): Unit = okTest {
-    val roleModel = initRoleModel("""
-                                    |{
-                                    |  "create-tables": [
-                                    |    {
-                                    |      "type": "grant",
-                                    |      "action": ["create"],
-                                    |      "scope": "table"
-                                    |    }
-                                    |  ]
-                                    |}""".stripMargin)
-
-    val controller = createStructureController(roleModel)
-
-    for {
-      permission <- controller.retrieveTables().map(getPermission)
-    } yield {
-
-      val expected = Json.obj(
-        "create" -> true
-      )
-
-      assertJSONEquals(expected, permission, JSONCompareMode.LENIENT)
-    }
-  }
-
-  @Test
-  def enrichTableSeq_createIsNotAllowed(implicit c: TestContext): Unit = okTest {
-    val controller = createStructureController()
-
-    for {
-      permission <- controller.retrieveTables().map(getPermission)
-    } yield {
-
-      val expected = Json.obj(
-        "create" -> false
-      )
-
-      assertJSONEquals(expected, permission, JSONCompareMode.LENIENT)
-    }
-  }
-
-  @Test
-  def enrichTable_editPropertiesAreAllowed(implicit c: TestContext): Unit = okTest {
-    val roleModel = initRoleModel("""
-                                    |{
-                                    |  "edit-tables": [
-                                    |    {
-                                    |      "type": "grant",
-                                    |      "action": ["view", "editDisplayProperty", "editStructureProperty"],
-                                    |      "scope": "table"
-                                    |    }
-                                    |  ]
-                                    |}""".stripMargin)
-
-    val controller = createStructureController(roleModel)
-
-    for {
-      tableId <- createDefaultTable()
-      permission <- controller.retrieveTable(tableId).map(getPermission)
-    } yield {
-
-      val expected = Json.obj(
-        "editDisplayProperty" -> true,
-        "editStructureProperty" -> true
-      )
-
-      assertJSONEquals(expected, permission, JSONCompareMode.LENIENT)
-    }
-  }
-
-  @Test
-  def enrichTable_editStructureProperty_onlyForModelTablesAllowed(implicit c: TestContext): Unit = okTest {
-    val roleModel = initRoleModel("""
-                                    |{
-                                    |  "edit-tables": [
-                                    |    {
-                                    |      "type": "grant",
-                                    |      "action": ["view", "editDisplayProperty"],
-                                    |      "scope": "table"
-                                    |    },
-                                    |    {
-                                    |      "type": "grant",
-                                    |      "action": ["editStructureProperty"],
-                                    |      "scope": "table",
-                                    |      "condition": {
-                                    |        "table": {
-                                    |          "name": ".*_model"
-                                    |        }
-                                    |      }
-                                    |    }
-                                    |  ]
-                                    |}""".stripMargin)
-
-    val controller = createStructureController(roleModel)
-
-    for {
-      modelTableId <- createDefaultTable("test_model", 1)
-      variantTableId <- createDefaultTable("test_variant", 2)
-      modelTablePermissions <- controller.retrieveTable(modelTableId).map(getPermission)
-      variantTablePermissions <- controller.retrieveTable(variantTableId).map(getPermission)
-    } yield {
-
-      val modelTableExpected = Json.obj(
-        "editDisplayProperty" -> true,
-        "editStructureProperty" -> true
-      )
-
-      val variantTableExpected = Json.obj(
-        "editDisplayProperty" -> true,
-        "editStructureProperty" -> false
-      )
-
-      assertJSONEquals(modelTableExpected, modelTablePermissions, JSONCompareMode.LENIENT)
-      assertJSONEquals(variantTableExpected, variantTablePermissions, JSONCompareMode.LENIENT)
-    }
-  }
-
-  @Test
-  def enrichTable_noActionIsAllowed(implicit c: TestContext): Unit = okTest {
-    val roleModel = initRoleModel("""
-                                    |{
-                                    |  "edit-tables": [
-                                    |    {
-                                    |      "type": "grant",
-                                    |      "action": ["view"],
-                                    |      "scope": "table"
-                                    |    }
-                                    |  ]
-                                    |}""".stripMargin)
-
-    val controller = createStructureController(roleModel)
-
-    for {
-      tableId <- createDefaultTable()
-      permission <- controller.retrieveTable(tableId).map(getPermission)
-    } yield {
-
-      val expected = Json.obj(
-        "editDisplayProperty" -> false,
-        "editStructureProperty" -> false,
-        "delete" -> false,
-        "createRow" -> false,
-        "deleteRow" -> false,
-        "editCellAnnotation" -> false,
-        "editRowAnnotation" -> false
-      )
-
-      assertJSONEquals(expected, permission, JSONCompareMode.LENIENT)
-    }
-  }
-
-  @Test
-  def enrichTable_allActionsAreAllowed(implicit c: TestContext): Unit = okTest {
-    val roleModel = initRoleModel("""
-                                    |{
-                                    |  "edit-tables": [
-                                    |    {
-                                    |      "type": "grant",
-                                    |      "action": ["view", "editDisplayProperty", "editStructureProperty", "delete",
-                                    |        "createRow", "deleteRow", "editCellAnnotation", "editRowAnnotation"],
-                                    |      "scope": "table"
-                                    |    },
-                                    |    {
-                                    |      "type": "grant",
-                                    |      "action": ["view", "create"],
-                                    |      "scope": "column"
-                                    |    }
-                                    |  ]
-                                    |}""".stripMargin)
-
-    val controller = createStructureController(roleModel)
-
-    for {
-      tableId <- createDefaultTable()
-      permission <- controller.retrieveTable(tableId).map(getPermission)
-    } yield {
-
-      val expected = Json.obj(
-        "editDisplayProperty" -> true,
-        "editStructureProperty" -> true,
-        "delete" -> true,
-        "createRow" -> true,
-        "deleteRow" -> true,
-        "editCellAnnotation" -> true,
-        "editRowAnnotation" -> true
-      )
-
-      assertJSONEquals(expected, permission, JSONCompareMode.LENIENT)
-    }
-  }
-
-  @Test
-  def enrichColumnSeq_createIsAllowed(implicit c: TestContext): Unit = okTest {
-    val roleModel = initRoleModel("""
-                                    |{
-                                    |  "create-columns": [
-                                    |    {
-                                    |      "type": "grant",
-                                    |      "action": ["view"],
-                                    |      "scope": "table"
-                                    |    },
-                                    |    {
-                                    |      "type": "grant",
-                                    |      "action": ["view", "create"],
-                                    |      "scope": "column"
-                                    |    }
-                                    |  ]
-                                    |}""".stripMargin)
-
-    val controller = createStructureController(roleModel)
-
-    for {
-      tableId <- createDefaultTable()
-      permission <- controller.retrieveColumns(tableId).map(getPermission)
-    } yield {
-      val expected = Json.obj(
-        "create" -> true
-      )
-
-      assertJSONEquals(expected, permission, JSONCompareMode.LENIENT)
-    }
-  }
-
-  @Test
-  def enrichColumnSeq_createIsNotAllowed(implicit c: TestContext): Unit = okTest {
-    val roleModel = initRoleModel("""
-                                    |{
-                                    |  "create-columns": [
-                                    |    {
-                                    |      "type": "grant",
-                                    |      "action": ["view"],
-                                    |      "scope": "table"
-                                    |    }
-                                    |  ]
-                                    |}""".stripMargin)
-
-    val controller = createStructureController(roleModel)
-
-    for {
-      tableId <- createDefaultTable()
-      permission <- controller.retrieveColumns(tableId).map(getPermission)
-    } yield {
-      val expected = Json.obj(
-        "create" -> false
-      )
-
-      assertJSONEquals(expected, permission, JSONCompareMode.LENIENT)
-    }
-  }
-
-  @Test
-  def enrichColumn_noActionIsAllowed(implicit c: TestContext): Unit = okTest {
-    val roleModel = initRoleModel("""
-                                    |{
-                                    |  "edit-columns": [
-                                    |    {
-                                    |      "type": "grant",
-                                    |      "action": ["view"],
-                                    |      "scope": "table"
-                                    |    },
-                                    |    {
-                                    |      "type": "grant",
-                                    |      "action": ["view"],
-                                    |      "scope": "column"
-                                    |    }
-                                    |  ]
-                                    |}""".stripMargin)
-
-    val controller = createStructureController(roleModel)
-
-    for {
-      tableId <- createDefaultTable()
-      permission <- controller.retrieveColumn(tableId, 1).map(getPermission)
-    } yield {
-      val expected = Json.obj(
-        "editDisplayProperty" -> false,
-        "editStructureProperty" -> false,
-        "delete" -> false,
-        "editCellValue" -> false
-      )
-
-      assertJSONEquals(expected, permission, JSONCompareMode.LENIENT)
-    }
-  }
-
-  @Test
-  def enrichColumn_allActionsAreAllowed(implicit c: TestContext): Unit = okTest {
-    val roleModel = initRoleModel("""
-                                    |{
-                                    |  "edit-columns": [
-                                    |    {
-                                    |      "type": "grant",
-                                    |      "action": ["view"],
-                                    |      "scope": "table"
-                                    |    },
-                                    |    {
-                                    |      "type": "grant",
-                                    |      "action": ["view", "editDisplayProperty", "editStructureProperty", "delete",
-                                    |        "editCellValue"],
-                                    |      "scope": "column"
-                                    |    }
-                                    |  ]
-                                    |}""".stripMargin)
-
-    val controller = createStructureController(roleModel)
-
-    for {
-      tableId <- createDefaultTable()
-      permission <- controller.retrieveColumn(tableId, 1).map(getPermission)
-    } yield {
-      val expected = Json.obj(
-        "editDisplayProperty" -> true,
-        "editStructureProperty" -> true,
-        "delete" -> true,
-        "editCellValue" -> true
-      )
-
-      assertJSONEquals(expected, permission, JSONCompareMode.LENIENT)
-    }
-  }
-
-  @Test
-  def enrichColumn_editCellValueIsAllowedForLangtagsDeAndEs(implicit c: TestContext): Unit = okTest {
-    val roleModel = initRoleModel("""
-                                    |{
-                                    |  "edit-columns": [
-                                    |    {
-                                    |      "type": "grant",
-                                    |      "action": ["view"],
-                                    |      "scope": "table"
-                                    |    },
-                                    |    {
-                                    |      "type": "grant",
-                                    |      "action": ["view", "editCellValue"],
-                                    |      "scope": "column",
-                                    |      "condition": {
-                                    |        "langtag": "en|es"
-                                    |      }
-                                    |    }
-                                    |  ]
-                                    |}""".stripMargin)
-
-    val controller = createStructureController(roleModel)
-
-    for {
-      _ <- sendRequest("POST", "/system/settings/langtags", Json.obj("value" -> Json.arr("de", "en", "fr", "es")))
-
-      _ <- createFullTableWithMultilanguageColumns("Test Table")
-      permission1 <- controller.retrieveColumn(1, 1).map(getPermission)
-      permission2 <- controller.retrieveColumn(1, 2).map(getPermission)
-    } yield {
-      val expected = Json.obj("de" -> false, "en" -> true, "fr" -> false, "es" -> true)
-
-      assertJSONEquals(expected, permission1.getJsonObject("editCellValue"), JSONCompareMode.LENIENT)
-      assertJSONEquals(expected, permission2.getJsonObject("editCellValue"), JSONCompareMode.LENIENT)
-    }
-  }
-
-  @Test
-  def enrichColumn_editCellValueWithoutLangtagConditionIsAllowedForAllLangtags(implicit c: TestContext): Unit = okTest {
-    val roleModel = initRoleModel("""
-                                    |{
-                                    |  "edit-columns": [
-                                    |    {
-                                    |      "type": "grant",
-                                    |      "action": ["view"],
-                                    |      "scope": "table"
-                                    |    },
-                                    |    {
-                                    |      "type": "grant",
-                                    |      "action": ["view", "editCellValue"],
-                                    |      "scope": "column"
-                                    |    }
-                                    |  ]
-                                    |}""".stripMargin)
-
-    val controller = createStructureController(roleModel)
-
-    for {
-      _ <- sendRequest("POST", "/system/settings/langtags", Json.obj("value" -> Json.arr("de", "en", "fr", "es")))
-
-      _ <- createFullTableWithMultilanguageColumns("Test Table")
-      permission <- controller.retrieveColumn(1, 1).map(getPermission)
-    } yield {
-      val expected = Json.obj("de" -> true, "en" -> true, "fr" -> true, "es" -> true)
-
-      assertJSONEquals(expected, permission.getJsonObject("editCellValue"), JSONCompareMode.LENIENT)
-    }
-  }
-
-  @Test
-  def enrichColumn_editCellValueWithoutLangtagConditionIsAllowedForAllCountryTags(implicit c: TestContext): Unit =
-    okTest {
-      val roleModel = initRoleModel("""
-                                      |{
-                                      |  "edit-columns": [
-                                      |    {
-                                      |      "type": "grant",
-                                      |      "action": ["view"],
-                                      |      "scope": "table"
-                                      |    },
-                                      |    {
-                                      |      "type": "grant",
-                                      |      "action": ["view", "editCellValue"],
-                                      |      "scope": "column"
-                                      |    }
-                                      |  ]
-                                      |}""".stripMargin)
-
-      val controller = createStructureController(roleModel)
-
-      val country_column = Json.obj(
-        "name" -> "country_column",
-        "kind" -> "currency",
-        "languageType" -> "country",
-        "countryCodes" -> Json.arr("DE", "AT", "GB")
-      )
-
-      val columns = Json.obj("columns" -> Json.arr(country_column))
-
-      for {
-        _ <- createEmptyDefaultTable()
-        _ <- sendRequest("POST", s"/tables/1/columns", columns)
-
-        permission <- controller.retrieveColumn(1, 3).map(getPermission)
-      } yield {
-        val expected = Json.obj("DE" -> true, "AT" -> true, "GB" -> true)
-        assertJSONEquals(expected, permission.getJsonObject("editCellValue"), JSONCompareMode.LENIENT)
-      }
-    }
-
-  @Test
-  def enrichColumn_editCellValueIsAllowedForCountryTagsATAndGB(implicit c: TestContext): Unit = okTest {
-    val roleModel = initRoleModel("""
-                                    |{
-                                    |  "edit-columns": [
-                                    |    {
-                                    |      "type": "grant",
-                                    |      "action": ["view"],
-                                    |      "scope": "table"
-                                    |    },
-                                    |    {
-                                    |      "type": "grant",
-                                    |      "action": ["view", "editCellValue"],
-                                    |      "scope": "column",
-                                    |      "condition": {
-                                    |        "langtag": "AT|GB"
-                                    |      }
-                                    |    }
-                                    |  ]
-                                    |}""".stripMargin)
-
-    val controller = createStructureController(roleModel)
-
-    val country_column = Json.obj(
-      "name" -> "country_column",
-      "kind" -> "currency",
-      "languageType" -> "country",
-      "countryCodes" -> Json.arr("DE", "AT", "GB")
-    )
-
-    val columns = Json.obj("columns" -> Json.arr(country_column))
-
-    for {
-      _ <- createEmptyDefaultTable()
-      _ <- sendRequest("POST", s"/tables/1/columns", columns)
-
-      permission <- controller.retrieveColumn(1, 3).map(getPermission)
-    } yield {
-      val expected = Json.obj("DE" -> false, "AT" -> true, "GB" -> true)
-      assertJSONEquals(expected, permission.getJsonObject("editCellValue"), JSONCompareMode.LENIENT)
-    }
-  }
-}
