+++ src/test/scala/com/campudus/tableaux/api/auth/SystemControllerAuthTest.scala
@@ -48,122 +48,6 @@ trait SystemControllerAuthTest extends TableauxTestBase {
   }
 }
 
-@RunWith(classOf[VertxUnitRunner])
-class SystemControllerAuthTest_enrichDomainObjects extends SystemControllerAuthTest {
-
-  @Test
-  def retrieveServices_createIsAllowed(implicit c: TestContext): Unit = okTest {
-    val roleModel = initRoleModel("""
-                                    |{
-                                    |  "view-create-services": [
-                                    |    {
-                                    |      "type": "grant",
-                                    |      "action": ["view", "create"],
-                                    |      "scope": "service"
-                                    |    }
-                                    |  ]
-                                    |}""".stripMargin)
-
-    val controller = createSystemController(roleModel)
-
-    for {
-      serviceId <- createDefaultService()
-      permission <- controller.retrieveServices().map(getPermission)
-    } yield {
-      val expected = Json.obj(
-        "create" -> true
-      )
-
-      assertJSONEquals(expected, permission, JSONCompareMode.STRICT)
-    }
-  }
-
-  @Test
-  def retrieveServices_createIsNotAllowed(implicit c: TestContext): Unit = okTest {
-    val roleModel = initRoleModel("""
-                                    |{
-                                    |  "view-services": [
-                                    |    {
-                                    |      "type": "grant",
-                                    |      "action": ["view"],
-                                    |      "scope": "service"
-                                    |    }
-                                    |  ]
-                                    |}""".stripMargin)
-
-    val controller = createSystemController(roleModel)
-
-    for {
-      serviceId <- createDefaultService()
-      permission <- controller.retrieveServices().map(getPermission)
-    } yield {
-      val expected = Json.obj(
-        "create" -> false
-      )
-
-      assertJSONEquals(expected, permission, JSONCompareMode.STRICT)
-    }
-  }
-
-  @Test
-  def retrieveService_allActionsAreAllowed(implicit c: TestContext): Unit = okTest {
-    val roleModel = initRoleModel("""
-                                    |{
-                                    |  "edit-and-delete-services": [
-                                    |    {
-                                    |      "type": "grant",
-                                    |      "action": ["view", "editDisplayProperty", "editStructureProperty", "delete"],
-                                    |      "scope": "service"
-                                    |    }
-                                    |  ]
-                                    |}""".stripMargin)
-
-    val controller = createSystemController(roleModel)
-
-    for {
-      serviceId <- createDefaultService()
-      permission <- controller.retrieveService(serviceId).map(getPermission)
-    } yield {
-      val expected = Json.obj(
-        "editDisplayProperty" -> true,
-        "editStructureProperty" -> true,
-        "delete" -> true
-      )
-
-      assertJSONEquals(expected, permission, JSONCompareMode.STRICT)
-    }
-  }
-
-  @Test
-  def retrieveService_noActionIsAllowed(implicit c: TestContext): Unit = okTest {
-    val roleModel = initRoleModel("""
-                                    |{
-                                    |  "view-services": [
-                                    |    {
-                                    |      "type": "grant",
-                                    |      "action": ["view"],
-                                    |      "scope": "service"
-                                    |    }
-                                    |  ]
-                                    |}""".stripMargin)
-
-    val controller = createSystemController(roleModel)
-
-    for {
-      serviceId <- createDefaultService()
-      permission <- controller.retrieveService(serviceId).map(getPermission)
-    } yield {
-      val expected = Json.obj(
-        "editDisplayProperty" -> false,
-        "editStructureProperty" -> false,
-        "delete" -> false
-      )
-
-      assertJSONEquals(expected, permission, JSONCompareMode.STRICT)
-    }
-  }
-
-}
 
 @RunWith(classOf[VertxUnitRunner])
 class SystemControllerAuthTest_checkAuthorization extends SystemControllerAuthTest {
