+++ src/test/scala/com/campudus/tableaux/api/media/MediaTest.scala
@@ -45,6 +45,128 @@ trait MediaTestBase extends TableauxTestBase {
   }
 }
 
+// @RunWith(classOf[VertxUnitRunner])
+// class MediaAuthTest extends MediaTestBase {
+
+//   def getPermission(domainObject: DomainObject): JsonObject = {
+//     domainObject.getJson.getJsonObject("permission")
+//   }
+
+//   @Test
+//   def enrich_noPermission_noActionIsAllowed(implicit c: TestContext): Unit =
+//     okTest {
+
+//       for {
+//         // permission <- controller.retrieveRootFolder("de").map(getPermission)
+//         permission <- sendRequest("GET", "/folders").map(getPermission)
+//       } yield {
+
+//         val expected = Json.obj(
+//           "create" -> false,
+//           "edit" -> false,
+//           "delete" -> false
+//         )
+
+//         assertJSONEquals(expected, permission)
+//       }
+//     }
+
+//   @Test
+//   def enrich_createGranted_onlyCreateIsAllowed(implicit c: TestContext): Unit = {
+//     val roleModel = initRoleModel("""
+//                                     |{
+//                                     |  "create-media": [
+//                                     |    {
+//                                     |      "type": "grant",
+//                                     |      "action": ["create"],
+//                                     |      "scope": "media"
+//                                     |    }
+//                                     |  ]
+//                                     |}""".stripMargin)
+
+
+//     okTest {
+//       for {
+//         permission <- controller.retrieveRootFolder("de").map(getPermission)
+//       } yield {
+
+//         val expected = Json.obj("create" -> true)
+
+//         assertJSONEquals(expected, permission, JSONCompareMode.LENIENT)
+//       }
+//     }
+//   }
+
+//   @Test
+//   def enrich_editAndDeleteGranted_onlyEditAndDeleteAreAllowed(implicit c: TestContext): Unit = {
+//     val roleModel = initRoleModel("""
+//                                     |{
+//                                     |  "edit-delete-media": [
+//                                     |    {
+//                                     |      "type": "grant",
+//                                     |      "action": ["edit", "delete"],
+//                                     |      "scope": "media"
+//                                     |    }
+//                                     |  ]
+//                                     |}""".stripMargin)
+
+
+//     okTest {
+//       for {
+//         permission <- controller.retrieveRootFolder("de").map(getPermission)
+//       } yield {
+
+//         val expected = Json.obj(
+//           "create" -> false,
+//           "edit" -> true,
+//           "delete" -> true
+//         )
+
+//         assertJSONEquals(expected, permission)
+//       }
+//     }
+//   }
+
+//   @Test
+//   def enrich_editAndDeleteGranted_deleteDenied_onlyEditIsAllowed(implicit c: TestContext): Unit = {
+//     val roleModel = initRoleModel("""
+//                                     |{
+//                                     |  "edit-delete-media": [
+//                                     |    {
+//                                     |      "type": "grant",
+//                                     |      "action": ["edit", "delete"],
+//                                     |      "scope": "media"
+//                                     |    }
+//                                     |  ],
+//                                     |  "NOT-delete-media": [
+//                                     |    {
+//                                     |      "type": "deny",
+//                                     |      "action": ["delete"],
+//                                     |      "scope": "media"
+//                                     |    }
+//                                     |  ]
+//                                     |}""".stripMargin)
+
+
+//     okTest {
+//       for {
+//         permission <- controller.retrieveRootFolder("de").map(getPermission)
+//       } yield {
+
+//         val expected = Json.obj(
+//           "create" -> false,
+//           "edit" -> true,
+//           "delete" -> false
+//         )
+
+//         assertJSONEquals(expected, permission)
+//       }
+//     }
+//   }
+
+
+// }
+
 @RunWith(classOf[VertxUnitRunner])
 class AttachmentTest extends MediaTestBase {
